requirements.txt
fastmcp>=2.8.1
python-docx
msoffcrypto-tool
docx2pdf
python-dotenv
uvicorn
starlette
azure-storage-blob
azure-identity